refactor(notifications): move push rendering and tap routing out of main.dart
main.dart carried the whole notification path: whether a background push
renders a local notification, the local push renderer itself, the tap-target
resolver, and the four navigators a tap can land on. None of it is
entrypoint work, and _routeNotificationTap had four call sites — three in
the startup coordinator and one in _DivineAppState — so the entrypoint was
acting as a shared library.

Move it to notifications/notification_tap_router.dart and make the tap
router public, since crossing the library boundary is now explicit.
handleFirebaseMessagingBackgroundMessage loses @visibleForTesting: it has a
real production caller across libraries now — the vm:entry-point handler,
which stays in main.dart as Firebase requires.

Their two test files move to test/notifications/, mirroring the source.

main.dart: 2224 -> 1911 lines.
